==> 7. VMI_Table1.R <==
library(tidyverse)
library(table1)
source("/Users/chenxinlei/Library/Mobile Documents/com~apple~CloudDocs/Projects/AKI/Code/VMI/0. VMI_Functions.R")

data = read.csv("/Users/chenxinlei/Library/Mobile Documents/com~apple~CloudDocs/Projects/AKI/Data/VMI/Table1_Complete_Data.csv") %>%
  select(-PatientID) %>%
  mutate(Sex_Scr = ifelse(Sex_Scr==1, "Female", "Male"),
         race = ifelse(race == 1, "White", ifelse(race == 2, "Black or African American", ifelse(race == 3, "Asian", "Other")))) %>%
  as.data.frame()


factor_var_list = data %>%
  summarise(across(everything(), ~ n_distinct(.))) %>%
  gather(variable, unique_count) %>%
  filter(unique_count < 8)

factor_var = c(
  "RRT_30days", "death_30days", "death_90days", "Sex_Scr", 
  "Vasopressin_1St_24Hrs_Yn", "race", "death", "ethnicity", 
  "residence_location", "lifestyle", 
  "comorbidities_mental", 
  "comorbidities_comat", "comorbidities_mvent", "comorbidities_chmvent", 
  "comorbidities_abrsmavail", "comorbidities_HTN", "comorbidities_mi", "comorbidities_achf", 
  "comorbidities_cad", "comorbidities_cerebral", "comorbidities_pvd", 
  "comorbidities_respdis", "comorbidities_resptype", 
  "comorbidities_dementia", "comorbidities_neuro", "comorbidities_endo", 
  "comorbidities_canc", "comorbidities_leukmm", "comorbidities_lymphoma", 
  "comorbidities_nonmettumor", "comorbidities_metcanc", 
  "comorbidities_immsupp", "comorbidities_aids", "comorbidities_renad", 
  "comorbidities_renadtype", "comorbidities_rendis", 
  "comorbidities_Ulcer", "comorbidities_cirrhos", 
  "comorbidities_portalHTN", "comorbidities_variceal", 
  "comorbidities_hepdis", "comorbidities_ochrol", 
  "comorbidities_musskel", "comorbidities_kidney", 
  "comorbidities_oralhypoglycemics", "comorbidities_aspirin7", 
  "comorbidities_steriods", "comorbidities_anticoags", 
  "comorbidities_statins", 
  "comorbidities_education",
  "AKI", "FIRST_AKI_DATE", 
  "FIRST_AKI_STAGE", "AKI_Category", 
  "AKI_Recovery", "AKI_Recovery_Total", "AKD", "MAKE30"
)


binary_list = c(
  "RRT_30days", "death_30days", "death_90days",
  "Vasopressin_1St_24Hrs_Yn", "death",
  "comorbidities_mental", 
  "comorbidities_comat", "comorbidities_mvent", "comorbidities_chmvent", 
  "comorbidities_abrsmavail", "comorbidities_HTN", "comorbidities_mi", "comorbidities_achf",  
  "comorbidities_cerebral", "comorbidities_pvd", 
  "comorbidities_respdis",
  "comorbidities_dementia", "comorbidities_neuro",
  "comorbidities_canc", "comorbidities_leukmm", "comorbidities_lymphoma", 
  "comorbidities_nonmettumor", "comorbidities_metcanc", 
  "comorbidities_immsupp", "comorbidities_aids", "comorbidities_renad", 
  "comorbidities_rendis", 
  "comorbidities_Ulcer", "comorbidities_cirrhos", 
  "comorbidities_portalHTN", "comorbidities_variceal", 
  "comorbidities_hepdis", "comorbidities_ochrol", 
  "comorbidities_musskel", "comorbidities_kidney", 
  "comorbidities_oralhypoglycemics", "comorbidities_aspirin7", 
  "comorbidities_steriods", "comorbidities_anticoags", 
  "comorbidities_statins", 
  "AKI", "AKI_Recovery", "AKI_Recovery_Total", "AKD", "MAKE30"
)

data$AKI_Category = factor(data$AKI_Category, levels = c("No AKI", "No Persistent Severe AKI", "Transient AKI", "Persistent Severe AKI"))

data[binary_list] <- lapply(data[binary_list], function(x) {
  ifelse(x == 1, "Yes", "No")
})

data = data %>%
  mutate(across(all_of(factor_var), as.factor))

labels = list(
  variables=list(RRT_30days = "RRT at day 30",
                 death_30days = "Death at day 30",
                 death_90days = "Death at day 90",
                 ref_Creatinine = "Reference creatinine",
                 APACHEIII_baseline = "APACHEIII",
                 Age_At_Enrollment = "Age at enrollment",
                 Sex_Scr = "Gender",
                 Charlson = "Charlson",
                 Platelet_Count6 = "Platelet count at 6h",
                 Platelet_Count24 = "Platelet count at 24h",
                 Platelet_Count72 = "Platelet count at 72h",
                 Vasopressin_1St_24Hrs_Yn = "Vasopressin",
                 Icu_Los = "ICU length of stay",
                 Hosp_LOS = "Hospital length of stay",
                 race = "Race",
                 death = "Death",
                 KIM1Val0 = "KIM1 at 0h",
                 NGALVal0 = "NGAL at 0h",
                 TIMP2_0 = "TIMP2 at 0h",
                 IGFBP7_0 = "IGFBP7 at 0h",
                 TNFVal6 = "TNFVal at 6h",
                 IL6Val6 = "IL6 at 6h",
                 IL10Val6 = "IL10 at 6h",
                 CRPVal6 = "CRP at 6h",
                 KIM1Val6 = "KIM1 at 6h",
                 NGALVal6 = "NGAL at 6h",
                 TNFVal24 = "TNF at 24h",
                 IL6Val24 = "IL6 at 24h",
                 IL10Val24 = "IL10 at 24h",
                 CRPVal24 = "CRP at 24h",
                 KIM1Val24 = "KIM1 at 24h",
                 NGALVal24 = "NGAL at 24h",
                 TNFVal72 = "TNF at 72h",
                 IL6Val72 = "IL6 at 72h",
                 IL10Val72 = "IL10 at 72h",
                 CRPVal72 = "CRP at 72h",
                 KIM1Val72 = "KIM1 at 72h",
                 NGALVal72 = "NGAL at 72h",
                 ethnicity = "Ethnicity",
                 weight = "Weight",
                 height = "Height",
                 residence_location = "Residence location",
                 employment_status = "Employment status",
                 lifestyle = "Lifestyle",
                 comorbidities_mental = "Comorbidities - mental",
                 comorbidities_comat = "Comorbidities - comat",
                 comorbidities_mvent = "Comorbidities - mvent",
                 comorbidities_chmvent = "Comorbidities - chmvent",
                 comorbidities_abrsmavail = "Comorbidities - abrsmavail",
                 comorbidities_abrsm = "Comorbidities - abrsm",
                 comorbidities_IVFvol = "Comorbidities - IVFvol",
                 comorbidities_education = "Comorbidities - education",
                 comorbidities_HTN = "Comorbidities - HTN",
                 comorbidities_mi = "Comorbidities - mi",
                 comorbidities_achf = "Comorbidities - achf",
                 comorbidities_cad = "Comorbidities - cad",
                 comorbidities_cerebral = "Comorbidities - cerebral",
                 comorbidities_pvd = "Comorbidities - pvd",
                 comorbidities_respdis = "Comorbidities - respdis",
                 comorbidities_resptype = "Comorbidities - resptype",
                 comorbidities_dementia = "Comorbidities - dementia",
                 comorbidities_neuro = "Comorbidities - neuro",
                 comorbidities_endo = "Comorbidities - endo",
                 comorbidities_canc = "Comorbidities - canc",
                 comorbidities_leukmm = "Comorbidities - leukmm",
                 comorbidities_lymphoma = "Comorbidities - lymphoma",
                 comorbidities_nonmettumor = "Comorbidities - nonmettumor",
                 comorbidities_metcanc = "Comorbidities - metcanc",
                 comorbidities_immsupp = "Comorbidities - immsupp",
                 comorbidities_aids = "Comorbidities - aids",
                 comorbidities_renad = "Comorbidities - renad",
                 comorbidities_renadtype = "Comorbidities - renadtype",
                 comorbidities_rendis = "Comorbidities - rendis",
                 comorbidities_Ulcer = "Comorbidities - Ulcer",
                 comorbidities_cirrhos = "Comorbidities - cirrhos",
                 comorbidities_portalHTN = "Comorbidities - portalHTN",
                 comorbidities_variceal = "Comorbidities - variceal",
                 comorbidities_hepdis = "Comorbidities - hepdis",
                 comorbidities_ochrol = "Comorbidities - ochrol",
                 comorbidities_musskel = "Comorbidities - musskel",
                 comorbidities_kidney = "Comorbidities - kidney",
                 comorbidities_oralhypoglycemics = "Comorbidities - oralhypoglycemics",
                 comorbidities_aspirin7 = "Comorbidities - aspirin7",
                 comorbidities_steriods = "Comorbidities - steroids",
                 comorbidities_anticoags = "Comorbidities - anticoags",
                 comorbidities_statins = "Comorbidities - statins",
                 X__24_hours_Total_Fluids = "Fluids within 24h",
                 total_HS_0 = "Total HS at 0h",
                 total_HS_6 = "Total HS at 6h",
                 total_HS_24 = "Total HS at 24h",
                 total_HS_72 = "Total HS at 72h",
                 PBR5_25_0 = "PBR 5-25 at 0h",
                 PBR5_25_6 = "PBR 5-25 at 6h",
                 PBR5_25_24 = "PBR 5-25 at 24h",
                 PBR5_25_72 = "PBR 5-25 at 72h",
                 PBR5_9_0 = "PBR 5-9 at 0h",
                 PBR5_9_6 = "PBR 5-9 at 6h",
                 PBR5_9_24 = "PBR 5-9 at 24h",
                 PBR5_9_72 = "PBR 5-9 at 72h",
                 PBR10_19_0 = "PBR 10-19 at 0h",
                 PBR10_19_6 = "PBR 10-19 at 6h",
                 PBR10_19_24 = "PBR 10-19 at 24h",
                 PBR10_19_72 = "PBR 10-19 at 72h",
                 PBR20_25_0 = "PBR 20-25 at 0h",
                 PBR20_25_6 = "PBR 20-25 at 6h",
                 PBR20_25_24 = "PBR 20-25 at 24h",
                 PBR20_25_72 = "PBR 20-25 at 72h",
                 Rolling_0 = "Rolling at 0h",
                 Rolling_6 = "Rolling at 6h",
                 Rolling_24 = "Rolling at 24h",
                 Rolling_72 = "Rolling at 72h",
                 Adhered_0 = "Adhesion at 0h",
                 Adhered_6 = "Adhesion at 6h",
                 Adhered_24 = "Adhesion at 24h",
                 Adhered_72 = "Adhesion at 72h",
                 DeBackerScore_0 = "DeBacker Score at 0h",
                 DeBackerScore_6 = "DeBacker Score at 6h",
                 DeBackerScore_24 = "DeBacker Score at 24h",
                 DeBackerScore_72 = "DeBacker Score at 72h",
                 SummaryTVD_Small_0 = "TVD at 0h",
                 SummaryTVD_Small_6 = "TVD at 6h",
                 SummaryTVD_Small_24 = "TVD at 24h",
                 SummaryTVD_Small_72 = "TVD at 72h",
                 SummaryPVD_Small_0 = "PVD at 0h",
                 SummaryPVD_Small_6 = "PVD at 6h",
                 SummaryPVD_Small_24 = "PVD at 24h",
                 SummaryPVD_Small_72 = "PVD at 72h",
                 SummaryPPV_Small_0 = "PPV at 0h",
                 SummaryPPV_Small_6 = "PPV at 6h",
                 SummaryPPV_Small_24 = "PPV at 24h",
                 SummaryPPV_Small_72 = "PPV at 72h",
                 HeteroIndexSmall_0 = "Heterogeneity Index at 0h",
                 HeteroIndexSmall_6 = "Heterogeneity Index at 6h",
                 HeteroIndexSmall_24 = "Heterogeneity Index at 24h",
                 HeteroIndexSmall_72 = "Heterogeneity Index at 72h",
                 MfiSmall_n_0 = "MFI at 0h",
                 MfiSmall_n_6 = "MFI at 6h",
                 MfiSmall_n_24 = "MFI at 24h",
                 MfiSmall_n_72 = "MFI at 72h",
                 ICAMVal_0 = "ICAM at 0h",
                 ICAMVal_6 = "ICAM at 6h",
                 ICAMVal_24 = "ICAM at 24h",
                 ICAMVal_72 = "ICAM at 72h",
                 VCAMVal_0 = "VCAM at 0h",
                 VCAMVal_6 = "VCAM at 6h",
                 VCAMVal_24 = "VCAM at 24h",
                 VCAMVal_72 = "VCAM at 72h",
                 E_SelectinVal_0 = "E selectin at 0h",
                 E_SelectinVal_6 = "E selectin at 6h",
                 E_SelectinVal_24 = "E selectin at 24h",
                 E_SelectinVal_72 = "E selectin at 72h",
                 P_SelectinVal_0 = "P selectin at 0h",
                 P_SelectinVal_6 = "P selectin at 6h",
                 P_SelectinVal_24 = "P selectin at 24h",
                 P_SelectinVal_72 = "P selectin at 72h",
                 Ang_2Val_0 = "Ang2 at 0h",
                 Ang_2Val_6 = "Ang2 at 6h",
                 Ang_2Val_24 = "Ang2 at 24h",
                 Ang_2Val_72 = "Ang2 at 72h",
                 VEGFVal_0 = "VEGF at 0h",
                 VEGFVal_6 = "VEGF at 0h",
                 VEGFVal_24 = "VEGF at 0h",
                 VEGFVal_72 = "VEGF at 0h",
                 AKI = "AKI",
                 FIRST_AKI_DATE = "First AKI date",
                 FIRST_AKI_STAGE = "First AKI stage",
                 AKI_Recovery = "AKI recovery",
                 AKI_Recovery_Total = "AKI recovery (Total)",
                 AKD = "AKD",
                 MAKE30 = "MAKE30"
                 ),
  groups=list("", "", "AKI"))

strata = c(list(Total=data), split(data, data$AKI_Category))
table1(~. | AKI_Category, overall=c(left="Total"), data=data)
table1(strata, labels, groupspan=c(1, 1, 3))




labels = list(
  variables=list(RRT_30days = "RRT at day 30",
                 death_30days = "Death at day 30",
                 death_90days = "Death at day 90",
                 ref_Creatinine = "Reference creatinine",
                 APACHEIII_baseline = "APACHEIII",
                 Age_At_Enrollment = "Age at enrollment",
                 Sex_Scr = "Gender",
                 Charlson = "Charlson",
                 Platelet_Count6 = "Platelet count at 6h",
                 Platelet_Count24 = "Platelet count at 24h",
                 Platelet_Count72 = "Platelet count at 72h",
                 Vasopressin_1St_24Hrs_Yn = "Vasopressin",
                 Icu_Los = "ICU length of stay",
                 Hosp_LOS = "Hospital length of stay",
                 race = "Race",
                 death = "Death",
                 KIM1Val0 = "KIM1 at 0h",
                 NGALVal0 = "NGAL at 0h",
                 TIMP2_0 = "TIMP2 at 0h",
                 IGFBP7_0 = "IGFBP7 at 0h",
                 TNFVal6 = "TNFVal at 6h",
                 IL6Val6 = "IL6 at 6h",
                 IL10Val6 = "IL10 at 6h",
                 CRPVal6 = "CRP at 6h",
                 KIM1Val6 = "KIM1 at 6h",
                 NGALVal6 = "NGAL at 6h",
                 TNFVal24 = "TNF at 24h",
                 IL6Val24 = "IL6 at 24h",
                 IL10Val24 = "IL10 at 24h",
                 CRPVal24 = "CRP at 24h",
                 KIM1Val24 = "KIM1 at 24h",
                 NGALVal24 = "NGAL at 24h",
                 TNFVal72 = "TNF at 72h",
                 IL6Val72 = "IL6 at 72h",
                 IL10Val72 = "IL10 at 72h",
                 CRPVal72 = "CRP at 72h",
                 KIM1Val72 = "KIM1 at 72h",
                 NGALVal72 = "NGAL at 72h",
                 ethnicity = "Ethnicity",
                 weight = "Weight",
                 height = "Height",
                 residence_location = "Residence location",
                 employment_status = "Employment status",
                 lifestyle = "Lifestyle",
                 comorbidities_mental = "Comorbidities - mental",
                 comorbidities_comat = "Comorbidities - comat",
                 comorbidities_mvent = "Comorbidities - mvent",
                 comorbidities_chmvent = "Comorbidities - chmvent",
                 comorbidities_abrsmavail = "Comorbidities - abrsmavail",
                 comorbidities_abrsm = "Comorbidities - abrsm",
                 comorbidities_IVFvol = "Comorbidities - IVFvol",
                 comorbidities_education = "Comorbidities - education",
                 comorbidities_HTN = "Comorbidities - HTN",
                 comorbidities_mi = "Comorbidities - mi",
                 comorbidities_achf = "Comorbidities - achf",
                 comorbidities_cad = "Comorbidities - cad",
                 comorbidities_cerebral = "Comorbidities - cerebral",
                 comorbidities_pvd = "Comorbidities - pvd",
                 comorbidities_respdis = "Comorbidities - respdis",
                 comorbidities_resptype = "Comorbidities - resptype",
                 comorbidities_dementia = "Comorbidities - dementia",
                 comorbidities_neuro = "Comorbidities - neuro",
                 comorbidities_endo = "Comorbidities - endo",
                 comorbidities_canc = "Comorbidities - canc",
                 comorbidities_leukmm = "Comorbidities - leukmm",
                 comorbidities_lymphoma = "Comorbidities - lymphoma",
                 comorbidities_nonmettumor = "Comorbidities - nonmettumor",
                 comorbidities_metcanc = "Comorbidities - metcanc",
                 comorbidities_immsupp = "Comorbidities - immsupp",
                 comorbidities_aids = "Comorbidities - aids",
                 comorbidities_renad = "Comorbidities - renad",
                 comorbidities_renadtype = "Comorbidities - renadtype",
                 comorbidities_rendis = "Comorbidities - rendis",
                 comorbidities_Ulcer = "Comorbidities - Ulcer",
                 comorbidities_cirrhos = "Comorbidities - cirrhos",
                 comorbidities_portalHTN = "Comorbidities - portalHTN",
                 comorbidities_variceal = "Comorbidities - variceal",
                 comorbidities_hepdis = "Comorbidities - hepdis",
                 comorbidities_ochrol = "Comorbidities - ochrol",
                 comorbidities_musskel = "Comorbidities - musskel",
                 comorbidities_kidney = "Comorbidities - kidney",
                 comorbidities_oralhypoglycemics = "Comorbidities - oralhypoglycemics",
                 comorbidities_aspirin7 = "Comorbidities - aspirin7",
                 comorbidities_steriods = "Comorbidities - steroids",
                 comorbidities_anticoags = "Comorbidities - anticoags",
                 comorbidities_statins = "Comorbidities - statins",
                 X__24_hours_Total_Fluids = "Fluids within 24h",
                 total_HS_0 = "Total HS at 0h",
                 total_HS_6 = "Total HS at 6h",
                 total_HS_24 = "Total HS at 24h",
                 total_HS_72 = "Total HS at 72h",
                 PBR5_25_0 = "PBR 5-25 at 0h",
                 PBR5_25_6 = "PBR 5-25 at 6h",
                 PBR5_25_24 = "PBR 5-25 at 24h",
                 PBR5_25_72 = "PBR 5-25 at 72h",
                 PBR5_9_0 = "PBR 5-9 at 0h",
                 PBR5_9_6 = "PBR 5-9 at 6h",
                 PBR5_9_24 = "PBR 5-9 at 24h",
                 PBR5_9_72 = "PBR 5-9 at 72h",
                 PBR10_19_0 = "PBR 10-19 at 0h",
                 PBR10_19_6 = "PBR 10-19 at 6h",
                 PBR10_19_24 = "PBR 10-19 at 24h",
                 PBR10_19_72 = "PBR 10-19 at 72h",
                 PBR20_25_0 = "PBR 20-25 at 0h",
                 PBR20_25_6 = "PBR 20-25 at 6h",
                 PBR20_25_24 = "PBR 20-25 at 24h",
                 PBR20_25_72 = "PBR 20-25 at 72h",
                 Rolling_0 = "Rolling at 0h",
                 Rolling_6 = "Rolling at 6h",
                 Rolling_24 = "Rolling at 24h",
                 Rolling_72 = "Rolling at 72h",
                 Adhered_0 = "Adhesion at 0h",
                 Adhered_6 = "Adhesion at 6h",
                 Adhered_24 = "Adhesion at 24h",
                 Adhered_72 = "Adhesion at 72h",
                 DeBackerScore_0 = "DeBacker Score at 0h",
                 DeBackerScore_6 = "DeBacker Score at 6h",
                 DeBackerScore_24 = "DeBacker Score at 24h",
                 DeBackerScore_72 = "DeBacker Score at 72h",
                 SummaryTVD_Small_0 = "TVD at 0h",
                 SummaryTVD_Small_6 = "TVD at 6h",
                 SummaryTVD_Small_24 = "TVD at 24h",
                 SummaryTVD_Small_72 = "TVD at 72h",
                 SummaryPVD_Small_0 = "PVD at 0h",
                 SummaryPVD_Small_6 = "PVD at 6h",
                 SummaryPVD_Small_24 = "PVD at 24h",
                 SummaryPVD_Small_72 = "PVD at 72h",
                 SummaryPPV_Small_0 = "PPV at 0h",
                 SummaryPPV_Small_6 = "PPV at 6h",
                 SummaryPPV_Small_24 = "PPV at 24h",
                 SummaryPPV_Small_72 = "PPV at 72h",
                 HeteroIndexSmall_0 = "Heterogeneity Index at 0h",
                 HeteroIndexSmall_6 = "Heterogeneity Index at 6h",
                 HeteroIndexSmall_24 = "Heterogeneity Index at 24h",
                 HeteroIndexSmall_72 = "Heterogeneity Index at 72h",
                 MfiSmall_n_0 = "MFI at 0h",
                 MfiSmall_n_6 = "MFI at 6h",
                 MfiSmall_n_24 = "MFI at 24h",
                 MfiSmall_n_72 = "MFI at 72h",
                 ICAMVal_0 = "ICAM at 0h",
                 ICAMVal_6 = "ICAM at 6h",
                 ICAMVal_24 = "ICAM at 24h",
                 ICAMVal_72 = "ICAM at 72h",
                 VCAMVal_0 = "VCAM at 0h",
                 VCAMVal_6 = "VCAM at 6h",
                 VCAMVal_24 = "VCAM at 24h",
                 VCAMVal_72 = "VCAM at 72h",
                 E_SelectinVal_0 = "E selectin at 0h",
                 E_SelectinVal_6 = "E selectin at 6h",
                 E_SelectinVal_24 = "E selectin at 24h",
                 E_SelectinVal_72 = "E selectin at 72h",
                 P_SelectinVal_0 = "P selectin at 0h",
                 P_SelectinVal_6 = "P selectin at 6h",
                 P_SelectinVal_24 = "P selectin at 24h",
                 P_SelectinVal_72 = "P selectin at 72h",
                 Ang_2Val_0 = "Ang2 at 0h",
                 Ang_2Val_6 = "Ang2 at 6h",
                 Ang_2Val_24 = "Ang2 at 24h",
                 Ang_2Val_72 = "Ang2 at 72h",
                 VEGFVal_0 = "VEGF at 0h",
                 VEGFVal_6 = "VEGF at 0h",
                 VEGFVal_24 = "VEGF at 0h",
                 VEGFVal_72 = "VEGF at 0h",
                 AKI = "AKI",
                 FIRST_AKI_DATE = "First AKI date",
                 FIRST_AKI_STAGE = "First AKI stage",
                 AKI_Recovery = "AKI recovery",
                 AKI_Recovery_Total = "AKI recovery (Total)",
                 AKD = "AKD",
                 MAKE30 = "MAKE30"
  ))

strata <- c(list(Total=data), split(data, data$AKD))
table1(~. | AKD, overall=c(left="Total"), data=data)
table1(strata, labels)
